+++ pydbml/constants.py
@@ -1,3 +1,4 @@
 ONE_TO_MANY = '<'
 MANY_TO_ONE = '>'
 ONE_TO_ONE = '-'
+MANY_TO_MANY = '<>'